Checking in
Fixed unstable item ID's based on file name and parsing order alone. Now the future saves will not break when blocks or biomes are added or altered.

+++ core/src/main/java/com/internal/bootstrap/worldpipeline/biome/BiomeHandle.java
@@ -7,13 +7,12 @@ public class BiomeHandle extends HandlePackage {
 
     // Identity
     private String biomeName;
-    private int biomeID;
+    private short biomeID;
     private Color biomeColor;
 
     // Constructor \\
 
-    public void constructor(String biomeName, int biomeID) {
-
+    public void constructor(String biomeName, short biomeID) {
         this.biomeName = biomeName;
         this.biomeID = biomeID;
         this.biomeColor = Color.WHITE;
@@ -25,7 +24,7 @@ public class BiomeHandle extends HandlePackage {
         return biomeName;
     }
 
-    public int getBiomeID() {
+    public short getBiomeID() {
         return biomeID;
     }
